+++ cohttp-eio/tests/test_chunk_server.ml
@@ -5,11 +5,11 @@ let dump_chunk buf chunk =
   let s = Format.asprintf "\n%a" Body.pp_chunk chunk in
   Buffer.add_string buf s
 
-let app (req, reader) =
+let app (req, reader, _client_addr) =
   match Http.Request.resource req with
   | "/" -> (
       let chunk_buf = Buffer.create 0 in
-      match Server.read_chunked (req, reader) (dump_chunk chunk_buf) with
+      match Server.read_chunked req reader (dump_chunk chunk_buf) with
       | headers ->
           let req = { req with headers } in
           Buffer.contents chunk_buf
@@ -24,5 +24,4 @@ let () =
     [ ("-p", Arg.Set_int port, " Listening port number(8080 by default)") ]
     ignore "An HTTP/1.1 server";
 
-  Eio_main.run @@ fun env ->
-  Eio.Switch.run @@ fun sw -> run ~port:!port env sw app
+  Eio_main.run @@ fun env -> run ~port:!port env app